backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.1
aiohttp==3.9.3
beautifulsoup4==4.12.3
anthropic>=0.77.0

# Slack Bot (JARVIS)
slack-bolt==1.18.1
slack-sdk==3.27.1

# File Processing
pandas==2.2.0
openpyxl==3.1.2
pypdf==4.0.1
python-docx==1.1.0

# Structured Logging
structlog==24.1.0
